html/semantics/embedded-content/the-area-element/area-coords.html

<!doctype html>
<meta charset=utf-8>
<title>HTMLAreaElement coords parsing</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
 body { margin: 0 }
</style>
<img src=/images/threecolors.png usemap=#x id=img width=300 height=300>
<map name=x><area id=area></map>
<script>
setup({explicit_done: true});

var hitRect = [[area, 3, 3], [area, 9, 9], [img, 1, 3], [img, 3, 1], [img, 11, 9], [img, 9, 11]];
var hitNone = [[img, 3, 3], [img, 9, 9], [img, 1, 3], [img, 3, 1], [img, 11, 9], [img, 9, 11]];

var tests = [
  {desc: 'sanity-check', shape: 'rect', coords: "2,2,10,10", hit: hitRect},
  {desc: 'SEMICOLON', shape: 'rect', coords: "2;2;10;10", hit: hitRect},
  {desc: 'SPACE', shape: 'rect', coords: "2 2 10 10", hit: hitRect},
  {desc: 'TAB', shape: 'rect', coords: "2\t2\t10\t10", hit: hitRect},
  {desc: 'FORM FEED', shape: 'rect', coords: "2\f2\f10\f10", hit: hitRect},
  {desc: 'LINE FEED', shape: 'rect', coords: "2\n2\n10\n10", hit: hitRect},
  {desc: 'CARRIGAGE RETURN', shape: 'rect', coords: "2\r2\r10\r10", hit: hitRect},
  {desc: 'LINE TABULATION', shape: 'rect', coords: "2\u000b2\u000b10\u000b10", hit: hitNone},
  {desc: 'LINE NEXT', shape: 'rect', coords: "2\u00852\u008510\u008510", hit: hitNone},
  {desc: 'EN QUAD', shape: 'rect', coords: "2\u20002\u200010\u200010", hit: hitNone},
  {desc: 'abc garbage between numbers', shape: 'rect', coords: "2a2b20c20,2,10,10", hit: hitRect},
  {desc: ': garbage between numbers', shape: 'rect', coords: "2:2:20:20,2,10,10", hit: hitRect},
  {desc: 'leading comma', shape: 'rect', coords: ",2,2,10,10", hit: hitRect},
  {desc: 'leading space', shape: 'rect', coords: " 2,2,10,10", hit: hitRect},
  {desc: 'leading semi-colon', shape: 'rect', coords: ";2,2,10,10", hit: hitRect},
  {desc: 'too few numbers', shape: 'rect', coords: "2,2,10", hit: hitNone},
  {desc: 'trailing comma', shape: 'rect', coords: "2,2,10,", hit: hitNone},
  {desc: 'trailing space', shape: 'rect', coords: "2,2,10 ", hit: hitNone},
  {desc: 'trailing semi-colon', shape: 'rect', coords: "2,2,10;", hit: hitNone},
  {desc: 'percent', shape: 'rect', coords: "2%,2%,10%,10%", hit: hitRect},
  {desc: 'float', shape: 'rect', coords: "1.4,1.4,10,10", hit: hitRect},
  {desc: 'number starting with dot', shape: 'rect', coords: ".4,.4,10,10", hit: [[area, 1, 1], [img, 0, 0]]},
  {desc: 'sci-not', shape: 'rect', coords: "2,2,1e1,1e1", hit: hitRect},
  {desc: 'leading/trailing garbage', shape: 'rect', coords: "='2,2,10,10' ", hit: hitRect},
  {desc: 'non-ascii garbage', shape: 'rect', coords: "“2,2,10,10\"", hit: hitRect},
  {desc: 'URL garbage with number', shape: 'rect', coords: "2,2,10ls/spain/holidays/regions/10/Canary+Islands/Canary+Islands.html", hit: hitNone},
  {desc: 'consecutive commas', shape: 'rect', coords: "2,,10,10", hit: hitNone},
  {desc: 'empty string', shape: 'rect', coords: "", hit: hitNone},
  // TODO(zcorpan): test other shapes
];
onload = function() {
  tests.forEach(function(t) {
    test(function(t_obj) {
      area.shape = t.shape;
      area.coords = t.coords;
      t.hit.forEach(function(arr) {
        var expected = arr[0];
        var x = arr[1];
        var y = arr[2];
        assert_equals(document.elementFromPoint(x, y), expected, 'elementFromPoint('+x+', '+y+')');
      });
    }, t.desc + ': ' + format_value(t.coords) + ' (' + t.shape + ')');
  });
  done();
};
</script>
